requirements.txt
python-telegram-bot==21.6
youtube-transcript-api==1.2.4
groq==0.11.0
httpx<0.28.0
python-dotenv==1.0.1
langdetect==1.0.9
pytest==8.3.3
pytest-asyncio==0.24.0
